requirements.txt
# Algae Microscopy Classifier - Requirements

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data loading and augmentation
albumentations>=1.3.0
pyyaml>=6.0

# Training and monitoring
tensorboard>=2.13.0
tqdm>=4.65.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Object detection (choose one based on your preference)
# Option 1: YOLOv8 (Recommended for this project)
ultralytics>=8.0.0

# Option 2: Detectron2 (Faster R-CNN, etc.)
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# Optional: For ONNX export
# onnx>=1.14.0
# onnxruntime>=1.15.0
